refactor: restructure services by domain
