--- notification-service/src/main/java/intexsoft/practice/notification_service/service/NotificationService.java
@@ -1,5 +1,0 @@
-package intexsoft.practice.notification_service.service;
-
-public interface NotificationService<T> {
-    void notify(T dto);
-}
