+++ notification-service/src/main/java/intexsoft/practice/notification_service/service/booking/BookingNotificationDataServiceImpl.java
@@ -1,8 +1,7 @@
-package intexsoft.practice.notification_service.service.impl;
+package intexsoft.practice.notification_service.service.booking;
 
 import intexsoft.practice.notification_service.entity.BookingNotification;
 import intexsoft.practice.notification_service.repository.BookingNotificationRepository;
-import intexsoft.practice.notification_service.service.BookingNotificationDataService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
