+++ notification-service/src/main/java/intexsoft/practice/notification_service/service/localization/LocaleMappingService.java
@@ -1,4 +1,4 @@
-package intexsoft.practice.notification_service.service;
+package intexsoft.practice.notification_service.service.localization;
 
 import java.util.Locale;
