+++ notification-service/src/main/java/intexsoft/practice/notification_service/service/mail/MailSenderService.java
@@ -1,4 +1,4 @@
-package intexsoft.practice.notification_service.service;
+package intexsoft.practice.notification_service.service.mail;
 
 public interface MailSenderService {
     void sendEmail(String recipientEmail, String subject, String body);
