+++ notification-service/src/main/java/intexsoft/practice/notification_service/service/notification/NotificationService.java
@@ -1,0 +1,5 @@
+package intexsoft.practice.notification_service.service.notification;
+
+public interface NotificationService<T> {
+    void notify(T dto);
+}
